+++ packages/cli/src/pages/_document.tsx
@@ -5,12 +5,12 @@ export default function Document() {
     <Html>
       <Head>
         <link
-          href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=optional"
+          href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=block"
           rel="stylesheet"
         />
         {/* Special characters in the text field... */}
         <link
-          href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=optional&text=%E2%97%8F;&#10003;&subset=latin"
+          href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=block&text=%E2%97%8F;&#10003;&subset=latin"
           rel="stylesheet"
         />
       </Head>