+++ packages/web/CHANGELOG.md
@@ -1,5 +1,31 @@
 # web
 
+## 0.1.2
+
+### Patch Changes
+
+- 6b969f7: font-display block
+- b940924: display block font-display
+- 44c6177: merge main with react17 support
+
+## 0.1.2-next.2
+
+### Patch Changes
+
+- font-display block
+
+## 0.1.2-next.1
+
+### Patch Changes
+
+- display block font-display
+
+## 0.1.2-next.0
+
+### Patch Changes
+
+- merge main with react17 support
+
 ## 0.1.1
 
 ### Patch Changes